Image Gallery Added
Created scroll view with paging for the gallery as well as folder structure for art resources in the project and added some temp gallery images to the project.

Known bugs:
If you goto the gallery view but then click on another tab, when you return there can be issues with the start position of the gallery and / or the alignment of the paging.  Not sure why.  Seems like if you scroll a direction before you leave and come back, it effects the view when you return.  If the bug happens you can scroll all the way to the beginning of the gallery again and then hit another tab and come back and it go's away (until you leave and return again…)