Design for the About Page
Added all of the art for the about page as well as some elements for the Estimate Page.  Also added a tint to the TabBar and fixed a bug on the iPhone Retina on the Estimate Page where the scroll frame was smaller than the device screen and cutting off content.